Use DocumentLocator in status popover rather than the document id.
