+++ packages/jupyterlab-lsp/src/features/diagnostics_listing.tsx
@@ -16,6 +16,7 @@ import { WidgetAdapter } from '../adapters/adapter';
 import { IVirtualEditor } from '../virtual/editor';
 import { CodeEditor } from '@jupyterlab/codeeditor';
 import { IDocumentWidget } from '@jupyterlab/docregistry';
+import { DocumentLocator, focus_on } from '../components/utils';
 
 export const diagnosticsIcon = new LabIcon({
   name: 'lsp:diagnostics',
@@ -44,81 +45,6 @@ export class DiagnosticsDatabase extends Map<
   get all(): Array<IEditorDiagnostic> {
     return [].concat.apply([], this.values());
   }
-}
-
-function focus_on(node: HTMLElement) {
-  if (!node) {
-    return;
-  }
-  node.scrollIntoView();
-  node.focus();
-}
-
-export function get_breadcrumbs(
-  document: VirtualDocument,
-  adapter: WidgetAdapter<IDocumentWidget>
-): JSX.Element[] {
-  return document.ancestry.map((document: VirtualDocument) => {
-    if (!document.parent) {
-      let path = document.path;
-      if (
-        !document.has_lsp_supported_file &&
-        path.endsWith(document.file_extension)
-      ) {
-        path = path.slice(0, -document.file_extension.length - 1);
-      }
-      return <span key={document.uri}>{path}</span>;
-    }
-    if (!document.virtual_lines.size) {
-      return <span key={document.uri}>Empty document</span>;
-    }
-    try {
-      if (adapter.has_multiple_editors) {
-        let first_line = document.virtual_lines.get(0);
-        let last_line = document.virtual_lines.get(
-          document.last_virtual_line - 1
-        );
-
-        let first_cell = adapter.get_editor_index(first_line.editor);
-        let last_cell = adapter.get_editor_index(last_line.editor);
-
-        let cell_locator =
-          first_cell === last_cell
-            ? `cell ${first_cell + 1}`
-            : `cells: ${first_cell + 1}-${last_cell + 1}`;
-
-        return (
-          <span key={document.uri}>
-            {document.language} ({cell_locator})
-          </span>
-        );
-      }
-    } catch (e) {
-      console.warn('LSP: could not display document cell location', e);
-    }
-    return <span key={document.uri}>{document.language}</span>;
-  });
-}
-
-function DocumentLocator(props: {
-  document: VirtualDocument;
-  adapter: WidgetAdapter<any>;
-}) {
-  let { document, adapter } = props;
-  let target: HTMLElement = null;
-  if (adapter.has_multiple_editors) {
-    let first_line = document.virtual_lines.get(0);
-    target = adapter.get_editor_wrapper(first_line.editor);
-  }
-  let breadcrumbs: any = get_breadcrumbs(document, adapter);
-  return (
-    <div
-      className={'lsp-document-locator'}
-      onClick={() => focus_on(target ? target : null)}
-    >
-      {breadcrumbs}
-    </div>
-  );
 }
 
 interface IDiagnosticsRow {
